requirements.txt
aws-cdk.core
aws-cdk.aws_apigateway
aws-cdk.aws_lambda
aws-cdk.aws_lambda_event_sources
aws-cdk.aws_dynamodb
aws-cdk.aws_events
aws-cdk.aws_events_targets

boto3
botocore